refactor: move block endpoints to new folder
