+++ crates/rpc/eth/block.rs
@@ -13,11 +13,8 @@ use crate::{
     utils::RpcErr,
     RpcHandler,
 };
-use ethereum_rust_core::{
-    rlp::encode::RLPEncode,
-    types::{
-        calculate_base_fee_per_blob_gas, Block, BlockBody, BlockHash, BlockHeader, BlockNumber,
-    },
+use ethereum_rust_core::types::{
+    calculate_base_fee_per_blob_gas, BlockBody, BlockHash, BlockHeader, BlockNumber,
 };
 use ethereum_rust_storage::{error::StoreError, Store};
 
@@ -37,10 +34,6 @@ pub struct GetBlockTransactionCountRequest {
 }
 
 pub struct GetBlockReceiptsRequest {
-    pub block: BlockIdentifierOrHash,
-}
-
-pub struct GetRawBlock {
     pub block: BlockIdentifierOrHash,
 }
 
@@ -196,39 +189,6 @@ impl RpcHandler for GetBlockReceiptsRequest {
         let receipts = get_all_block_receipts(block_number, header, body, &storage)?;
 
         serde_json::to_value(&receipts).map_err(|_| RpcErr::Internal)
-    }
-}
-
-impl RpcHandler for GetRawBlock {
-    fn parse(params: &Option<Vec<Value>>) -> Option<GetRawBlock> {
-        let params = params.as_ref()?;
-        if params.len() != 1 {
-            return None;
-        }
-        Some(GetRawBlock {
-            block: serde_json::from_value(params[0].clone()).ok()?,
-        })
-    }
-
-    fn handle(&self, storage: Store) -> Result<Value, RpcErr> {
-        info!("Requested raw block: {}", self.block);
-        let block_number = match self.block.resolve_block_number(&storage)? {
-            Some(block_number) => block_number,
-            _ => return Ok(Value::Null),
-        };
-        let header = storage.get_block_header(block_number)?;
-        let body = storage.get_block_body(block_number)?;
-        let (header, body) = match (header, body) {
-            (Some(header), Some(body)) => (header, body),
-            _ => return Ok(Value::Null),
-        };
-        let block = Block {
-            header: header.clone(),
-            body: body.clone(),
-        }
-        .encode_to_vec();
-
-        serde_json::to_value(format!("0x{}", &hex::encode(block))).map_err(|_| RpcErr::Internal)
     }
 }
